+++ roxcomposer/tests/base_service.py
@@ -113,6 +113,27 @@ class TestBaseService(unittest.TestCase):
 
         self.assertRaises(exceptions.ParameterMissing, base_service.BaseService, self.test_params_2)
 
+        # test invalid parameters
+        self.assertRaises(exceptions.ConfigError, base_service.BaseService, {
+            'ip': '127.0.0.1',
+            'port': '4711',
+            'name': 'fancy-service',
+            "logging": {
+                "level": "INFO",
+                "logpath": "/dev/null"
+            }
+        })
+
+        self.assertRaises(exceptions.ConfigError, base_service.BaseService, {
+            'ip': '127.0.999.1',
+            'port': 4711,
+            'name': 'fancy-service',
+            "logging": {
+                "level": "INFO",
+                "logpath": "/dev/null"
+            }
+        })
+
         # test initiation with params
         bs_with_params = base_service.BaseService({
             'ip': '127.0.0.1',
@@ -207,7 +228,7 @@ class TestBaseService(unittest.TestCase):
     @unittest.skipIf('SKIP_TEMPDIR_TEST' in os.environ, "tempdir issues")
     def test_config_loading(self):
         os.environ['DROXIT_ROXCOMPOSER_CONFIG'] = '/bogus/path/from/hell.json'
-        self.assertRaises(exceptions.ConfigError, base_service.BaseService, { "service_key": "nevermind" })
+        self.assertRaises(exceptions.ConfigError, base_service.BaseService, {"service_key": "nevermind"})
 
         with TemporaryDirectory() as tdir:
             confname = os.path.join(tdir, "config.json")
@@ -220,14 +241,14 @@ class TestBaseService(unittest.TestCase):
 
             os.environ['DROXIT_ROXCOMPOSER_CONFIG'] = confname
 
-            self.assertRaises(exceptions.ParameterMissing, base_service.BaseService, { "service_key": "nevermind" })
+            self.assertRaises(exceptions.ParameterMissing, base_service.BaseService, {"service_key": "nevermind"})
             s = base_service.BaseService({"service_key": "service.dummy"})
             self.assertDictEqual(s.params, self.test_params_1)
 
             f = open(confname, "w")
             f.write('this is no proper JSON')
             f.close()
-            self.assertRaises(exceptions.ConfigError, base_service.BaseService, { "service_key": "nevermind" })
+            self.assertRaises(exceptions.ConfigError, base_service.BaseService, {"service_key": "nevermind"})
 
 
 if __name__ == '__main__':